Extract translation classes from grampslocale.py
For name clarity and to reduce the line count in grampslocale.py that
Pylint thinks is excessive.
