+++ po/cs.po
@@ -2274,7 +2274,7 @@ msgid "{long_month} {year}"
 msgstr "{long_month} {year}"
 
 #. Translators: If "from <Month>" needs a special inflection
-#. in your language, translate to "{long_month.f[X]} {year}"
+#. in your language, translate to "{long_month.forms[X]} {year}"
 #. (where X is one of the month-name inflections you defined)
 #. else leave it untranslated
 #: ../gramps/gen/datehandler/_datedisplay.py:188
@@ -2284,7 +2284,7 @@ msgid "{long_month} {year}"
 msgstr "{long_month} {year}"
 
 #. Translators: If "to <Month>" needs a special inflection
-#. in your language, translate to "{long_month.f[X]} {year}"
+#. in your language, translate to "{long_month.forms[X]} {year}"
 #. (where X is one of the month-name inflections you defined)
 #. else leave it untranslated
 #: ../gramps/gen/datehandler/_datedisplay.py:195
@@ -2294,67 +2294,67 @@ msgid "{long_month} {year}"
 msgstr "{long_month} {year}"
 
 #. Translators: If "between <Month>" needs a special inflection
-#. in your language, translate to "{long_month.f[X]} {year}"
+#. in your language, translate to "{long_month.forms[X]} {year}"
 #. (where X is one of the month-name inflections you defined)
 #. else leave it untranslated
 #: ../gramps/gen/datehandler/_datedisplay.py:202
 #, python-brace-format
 msgctxt "between"
 msgid "{long_month} {year}"
-msgstr "{long_month.f[X]} {year}"
+msgstr "{long_month.forms[X]} {year}"
 
 #. Translators: If "and <Month>" needs a special inflection
-#. in your language, translate to "{long_month.f[X]} {year}"
+#. in your language, translate to "{long_month.forms[X]} {year}"
 #. (where X is one of the month-name inflections you defined)
 #. else leave it untranslated
 #: ../gramps/gen/datehandler/_datedisplay.py:209
 #, python-brace-format
 msgctxt "and"
 msgid "{long_month} {year}"
-msgstr "{long_month.f[O]} {year}"
+msgstr "{long_month.forms[O]} {year}"
 
 #. Translators: If "before <Month>" needs a special inflection
-#. in your language, translate to "{long_month.f[X]} {year}"
+#. in your language, translate to "{long_month.forms[X]} {year}"
 #. (where X is one of the month-name inflections you defined)
 #. else leave it untranslated
 #: ../gramps/gen/datehandler/_datedisplay.py:215
 #, python-brace-format
 msgctxt "before"
 msgid "{long_month} {year}"
-msgstr "{long_month.f[O]} {year}"
+msgstr "{long_month.forms[O]} {year}"
 
 #. Translators: If "after <Month>" needs a special inflection
-#. in your language, translate to "{long_month.f[X]} {year}"
+#. in your language, translate to "{long_month.forms[X]} {year}"
 #. (where X is one of the month-name inflections you defined)
 #. else leave it untranslated
 #: ../gramps/gen/datehandler/_datedisplay.py:221
 #, python-brace-format
 msgctxt "after"
 msgid "{long_month} {year}"
-msgstr "{long_month.f[T]} {year}"
+msgstr "{long_month.forms[T]} {year}"
 
 #. Translators: If "about <Month>" needs a special inflection
-#. in your language, translate to "{long_month.f[X]} {year}"
+#. in your language, translate to "{long_month.forms[X]} {year}"
 #. (where X is one of the month-name inflections you defined)
 #. else leave it untranslated
 #: ../gramps/gen/datehandler/_datedisplay.py:227
 #, python-brace-format
 msgctxt "about"
 msgid "{long_month} {year}"
-msgstr "{long_month.f[D]} {year}"
+msgstr "{long_month.forms[D]} {year}"
 
 #. Translators: If "estimated <Month>" needs a special inflection
-#. in your language, translate to "{long_month.f[X]} {year}"
+#. in your language, translate to "{long_month.forms[X]} {year}"
 #. (where X is one of the month-name inflections you defined)
 #. else leave it untranslated
 #: ../gramps/gen/datehandler/_datedisplay.py:233
 #, python-brace-format
 msgctxt "estimated"
 msgid "{long_month} {year}"
-msgstr "{long_month.f[D]} {year}"
+msgstr "{long_month.forms[D]} {year}"
 
 #. Translators: If "calculated <Month>" needs a special inflection
-#. in your language, translate to "{long_month.f[X]} {year}"
+#. in your language, translate to "{long_month.forms[X]} {year}"
 #. (where X is one of the month-name inflections you defined)
 #. else leave it untranslated
 #: ../gramps/gen/datehandler/_datedisplay.py:239
@@ -2369,7 +2369,7 @@ msgid "{short_month} {year}"
 msgstr "{short_month} {year}"
 
 #. Translators: If "from <Month>" needs a special inflection
-#. in your language, translate to "{short_month.f[X]} {year}"
+#. in your language, translate to "{short_month.forms[X]} {year}"
 #. (where X is one of the month-name inflections you defined)
 #. else leave it untranslated
 #: ../gramps/gen/datehandler/_datedisplay.py:250
@@ -2379,7 +2379,7 @@ msgid "{short_month} {year}"
 msgstr "{short_month} {year}"
 
 #. Translators: If "to <Month>" needs a special inflection
-#. in your language, translate to "{short_month.f[X]} {year}"
+#. in your language, translate to "{short_month.forms[X]} {year}"
 #. (where X is one of the month-name inflections you defined)
 #. else leave it untranslated
 #: ../gramps/gen/datehandler/_datedisplay.py:257
@@ -2389,7 +2389,7 @@ msgid "{short_month} {year}"
 msgstr "{short_month} {year}"
 
 #. Translators: If "between <Month>" needs a special inflection
-#. in your language, translate to "{short_month.f[X]} {year}"
+#. in your language, translate to "{short_month.forms[X]} {year}"
 #. (where X is one of the month-name inflections you defined)
 #. else leave it untranslated
 #: ../gramps/gen/datehandler/_datedisplay.py:264
@@ -2399,7 +2399,7 @@ msgid "{short_month} {year}"
 msgstr "{short_month} {year}"
 
 #. Translators: If "and <Month>" needs a special inflection
-#. in your language, translate to "{short_month.f[X]} {year}"
+#. in your language, translate to "{short_month.forms[X]} {year}"
 #. (where X is one of the month-name inflections you defined)
 #. else leave it untranslated
 #: ../gramps/gen/datehandler/_datedisplay.py:271
@@ -2409,7 +2409,7 @@ msgid "{short_month} {year}"
 msgstr "{short_month} {year}"
 
 #. Translators: If "before <Month>" needs a special inflection
-#. in your language, translate to "{short_month.f[X]} {year}"
+#. in your language, translate to "{short_month.forms[X]} {year}"
 #. (where X is one of the month-name inflections you defined)
 #. else leave it untranslated
 #: ../gramps/gen/datehandler/_datedisplay.py:277
@@ -2419,7 +2419,7 @@ msgid "{short_month} {year}"
 msgstr "{short_month} {year}"
 
 #. Translators: If "after <Month>" needs a special inflection
-#. in your language, translate to "{short_month.f[X]} {year}"
+#. in your language, translate to "{short_month.forms[X]} {year}"
 #. (where X is one of the month-name inflections you defined)
 #. else leave it untranslated
 #: ../gramps/gen/datehandler/_datedisplay.py:283
@@ -2429,7 +2429,7 @@ msgid "{short_month} {year}"
 msgstr "{short_month} {year}"
 
 #. Translators: If "about <Month>" needs a special inflection
-#. in your language, translate to "{short_month.f[X]} {year}"
+#. in your language, translate to "{short_month.forms[X]} {year}"
 #. (where X is one of the month-name inflections you defined)
 #. else leave it untranslated
 #: ../gramps/gen/datehandler/_datedisplay.py:289
@@ -2439,7 +2439,7 @@ msgid "{short_month} {year}"
 msgstr "{short_month} {year}"
 
 #. Translators: If "estimated <Month>" needs a special inflection
-#. in your language, translate to "{short_month.f[X]} {year}"
+#. in your language, translate to "{short_month.forms[X]} {year}"
 #. (where X is one of the month-name inflections you defined)
 #. else leave it untranslated
 #: ../gramps/gen/datehandler/_datedisplay.py:295
@@ -2449,7 +2449,7 @@ msgid "{short_month} {year}"
 msgstr "{short_month} {year}"
 
 #. Translators: If "calculated <Month>" needs a special inflection
-#. in your language, translate to "{short_month.f[X]} {year}"
+#. in your language, translate to "{short_month.forms[X]} {year}"
 #. (where X is one of the month-name inflections you defined)
 #. else leave it untranslated
 #: ../gramps/gen/datehandler/_datedisplay.py:301
